+++ packages/keyring-controller/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING:** Use new `Messenger` from `@metamask/messenger` ([#6370](https://github.com/MetaMask/core/pull/6370))
+  - Previously, `KeyringController` accepted a `RestrictedMessenger` instance from `@metamask/base-controller`.
+
 ## [23.1.1]
 
 ### Changed